Renamed src/models to src/common. Renamed tests/models to tests/mocks. Started work on logging and commenting
