+++ nodejs12.x/cookiecutter-typescript-app-template/{{cookiecutter.project_name}}/src/apps/lambda-app.ts
@@ -1,0 +1,6 @@
+import { ApiGatewayEvent } from '../common/apigateway/apigateway-event';
+import { ApiGatewayResponse } from '../common/apigateway/apigateway-response';
+
+export interface LambdaApp {
+    run(event: ApiGatewayEvent): Promise<ApiGatewayResponse>;
+}
