+++ nodejs12.x/cookiecutter-typescript-app-template/{{cookiecutter.project_name}}/src/common/apigateway/apigateway-request-context-identity.ts
