+++ nodejs12.x/cookiecutter-typescript-app-template/{{cookiecutter.project_name}}/src/common/todo-item.ts
